refactor p2 - touched some of the party logic - host switching
