+++ Play/app/src/main/java/com/miloshzelembaba/play/api/Request.java
@@ -7,34 +7,15 @@ import org.json.JSONObject;
 
 import java.util.HashMap;
 
-/**
- * Created by miloshzelembaba on 2018-03-07.
- */
-
 public class Request {
     private HashMap<String, Object> params;
     private String url;
 
     public Request(){
-        url = "https://shareplay-204722.appspot.com/";
-//        url = "http://10.0.0.15:8000/";
-//        if (ApplicationUtil.getInstance().getAppMode().equals("dev_server")) {
-//            url = "https://shareplay-204722.appspot.com/";
-//        } else if (ApplicationUtil.getInstance().getAppMode().equals("local")) {
-//            if (ApplicationUtil.getInstance().getLocalServerAddress().isEmpty()) {
-//                findLocalServer();
-//            }
-//            url = ApplicationUtil.getInstance().getLocalServerAddress();
-//        }
+//        url = "https://shareplay-204722.appspot.com/";
+        url = "http://10.0.0.15:8000/";
         params = new HashMap<>();
         addParameter("user", ApplicationUtil.getInstance().getUser());
-//        addParameter("request_id", System.currentTimeMillis());
-//        NetworkController.getInstance().put(Long.toString((Long)params.get("request_id")));
-    }
-
-    private void findLocalServer() {
-        ApplicationUtil.getInstance().setLocalServerAddress("http://192.168.0.21:8000/");
-        //            url = "http://192.168.0.21:8000/";
     }
 
     public void setUrl(String url){
@@ -47,10 +28,6 @@ public class Request {
             params.put(key, value);
         }
     }
-
-//    public Object getParameter(String id) {
-//        return params.get(id);
-//    }
 
     public String buildURL(){
         return url;
